FINAL RELEASE:
Completely refactored code. Main code in contest1.cpp has the bare
minimum. All other functionality is stored in ./utils header files.
CMakeLists.txt is updated accordingly.
