+++ src/utils/bumper.h
@@ -1,0 +1,10 @@
+#ifndef __BUMPER_H_INCLUDED__
+#define __BUMPER_H_INCLUDED__
+
+#include "contest1.h"
+
+// ==== BUMPER FUNCTION DECLARATIONS ====== 
+extern bool bumpersPressed();
+extern uint8_t specificBumperPressed();
+
+#endif
