+++ src/utils/wallFollow.h
@@ -1,0 +1,9 @@
+#ifndef __WALL_FOLLOW_H_INCLUDED__
+#define __WALL_FOLLOW_H_INCLUDED__
+
+#include "contest1.h"
+
+// ==== WALL FOLLOW FUNCTION DECLARATIONS ====== 
+extern void wallFollow(ros::Publisher *vel_pub);
+
+#endif
